requirements.txt
streamlit
matplotlib
seaborn
pandas
mysql-connector-python